# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DraftInput": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "expected_version",
          "payload"
        ],
        "title": "DraftInput",
        "type": "object"
      },
...
      "ExecuteInput": {
        "additionalProperties": false,
        "properties": {
          "confirmation": {
            "const": "REVOKE_CUSTOMER_ACCESS_KEEP_DATA",
            "title": "Confirmation",
            "type": "string"
          },
          "expected_digest": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Expected Digest",
            "type": "string"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "string"
          }
        },
        "required": [
          "plan_id",
          "expected_digest",
          "confirmation"
        ],
        "title": "ExecuteInput",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LifecycleRequest": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "data_export",
              "tenant_offboard"
            ],
            "title": "Kind",
            "type": "string"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 5,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "reason"
        ],
        "title": "LifecycleRequest",
        "type": "object"
      },
...
      "PlanInput": {
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "PlanInput",
        "type": "object"
      },
...
      "RetentionInput": {
        "additionalProperties": false,
        "properties": {
          "backup_expiry": {
            "format": "date-time",
            "title": "Backup Expiry",
            "type": "string"
          },
          "days": {
            "maximum": 36500.0,
            "minimum": 0.0,
            "title": "Days",
            "type": "integer"
          },
          "financial_settlement_confirmed": {
            "const": true,
            "title": "Financial Settlement Confirmed",
            "type": "boolean"
          },
          "legal_hold_cleared": {
            "const": true,
            "title": "Legal Hold Cleared",
            "type": "boolean"
          },
          "reference": {
            "maxLength": 500,
            "minLength": 8,
            "title": "Reference",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "days",
          "backup_expiry",
          "reference",
          "legal_hold_cleared",
          "financial_settlement_confirmed"
        ],
        "title": "RetentionInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Private security service platform. Development providers explicitly labeled; production release gate closed.",
    "title": "KUANGUARD API",
    "version": "0.1.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/app/drafts/{kind}/{key}": {
      "delete": {
        "operationId": "delete_draft_app_drafts__kind___key__delete",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "campaign",
                "training",
                "purchase"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Draft",
        "tags": [
          "drafts"
        ]
      },
      "get": {
        "operationId": "get_draft_app_drafts__kind___key__get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "campaign",
                "training",
                "purchase"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Draft",
        "tags": [
          "drafts"
        ]
      },
      "put": {
        "operationId": "save_draft_app_drafts__kind___key__put",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "campaign",
                "training",
                "purchase"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Draft",
        "tags": [
          "drafts"
        ]
      }
    },
...
    "/customer/data-export": {
      "get": {
        "operationId": "customer_export_customer_data_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "section",
            "required": true,
            "schema": {
              "enum": [
                "projects",
                "reports",
                "findings",
                "orders",
                "training",
                "campaigns",
                "tickets"
              ],
              "title": "Section",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Customer Export",
        "tags": [
          "data lifecycle"
        ]
      }
    },
...
    "/customer/lifecycle/requests": {
      "get": {
        "operationId": "customer_requests_customer_lifecycle_requests_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Customer Requests",
        "tags": [
          "data lifecycle"
        ]
      },
      "post": {
        "operationId": "create_request_customer_lifecycle_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LifecycleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Request",
        "tags": [
          "data lifecycle"
        ]
      }
    },
...
    "/internal/lifecycle/offboard": {
      "post": {
        "operationId": "execute_offboard_internal_lifecycle_offboard_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Offboard",
        "tags": [
          "data lifecycle"
        ]
      }
    },
    "/internal/lifecycle/plan": {
      "post": {
        "operationId": "create_plan_internal_lifecycle_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Plan",
        "tags": [
          "data lifecycle"
        ]
      }
    },
    "/internal/lifecycle/retention": {
      "post": {
        "operationId": "approve_retention_internal_lifecycle_retention_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Retention",
        "tags": [
          "data lifecycle"
        ]
      }
    },
...
  }
}